feat: added support for variables && info data enrichment && new parser method

+++ src/main/java/berlin/yuna/apidoccrafter/util/FileCleaner.java
@@ -16,15 +16,16 @@ public class FileCleaner {
             final Path tempDirectory = Path.of(System.getProperty("user.home"), "api-doc-cleaner");
             mkdir(tempDirectory);
             final Path target = tempDirectory.resolve(path.getFileName());
+            if (Files.exists(target))
+                return target;
             List<String> content = Files.readAllLines(path, getEncoding())
                 .stream()
                 .filter(line -> !line.matches("^\\s*(//|#).*"))
                 .filter(line -> !line.equals("---"))
-                .map(line -> line.equals("swagger: \"2.0\"")? "openapi: 3.0.1" : line)
-                .map(line -> line.equals("\"swagger\": \"2.0\"")? "\"openapi\": \"3.0.1\"" : line)
+                .map(line -> line.equals("swagger: \"2.0\"") ? "openapi: 3.0.1" : line)
+                .map(line -> line.equals("\"swagger\": \"2.0\"") ? "\"openapi\": \"3.0.1\"" : line)
                 .toList();
             Files.write(target, content, getEncoding());
-            System.out.println("[WARN] Cleaning [" + target + "] and retry parsing");
             return target;
         } catch (Exception e) {
             System.err.println("[ERROR] Failed to preprocess file [" + path + "] cause [" + e.getClass().getSimpleName() + "] message [" + e.getMessage() + "]");
